requirements.txt
alabaster==0.7.12
anyio==3.4.0
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.9.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
black==21.12b0
bleach==4.1.0
boto3==1.20.26
botocore==1.23.26
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.3
comet-git-pure==0.19.16
comet-ml==3.2.0
configobj==5.0.6
cryptography==36.0.0
cycler==0.11.0
Cython==0.29.26
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
docutils==0.17.1
entrypoints==0.3
everett==1.0.2
flake8==4.0.1
flake8-docstrings==1.6.0
flake8-import-order==0.18.1
Flask==2.0.2
Flask-Cors==3.0.10
fonttools==4.28.5
gmplot==1.4.1
idna==3.3
imageio==2.13.4
imagesize==1.3.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
json5==0.9.6
jsonpatch==1.32
jsonpointer==2.2
jsonschema==4.3.2
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.1
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
motmetrics==1.2.0
munkres==1.1.4
mypy==0.930
mypy-extensions==0.4.3
nanoid==2.0.0
nb-conda-kernels==2.3.1
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.9
networkx==2.6.3
notebook==6.4.6
numpy==1.21.5
nvidia-ml-py3==7.352.0
olefile==0.46
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.3.1
platformdirs==2.4.0
pluggy==1.0.0
plyfile==0.7.4
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
py-cpuinfo==8.0.0
pycocotools==2.0.3
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.10.0
pyhumps==3.0.2
pylint==2.12.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
pytest==6.2.5
pytest-benchmark==3.4.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
s3transfer==0.5.0
scalabel==0.3.0rc2
scikit-image==0.19.1
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==60.1.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
snowballstemmer==2.2.0
Sphinx==4.3.2
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tabulate==0.8.9
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.3
torch==1.10.1
torchaudio==0.10.1
torchfile==0.1.0
torchinfo==1.6.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typed-ast==1.5.1
typing_extensions==4.0.1
unicodedata2==14.0.0
urllib3==1.26.7
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
wheel==0.37.1
widgetsnbextension==3.5.2
wrapt==1.13.3
wurlitzer==1.0.3
xmltodict==0.12.0
zipp==3.6.0
